refactor: useNotes 커스텀 훅에 무한 스크롤 로직 통합
