+++ src/pages/notes/[goalId]/index.tsx
@@ -1,8 +1,6 @@
 import { useRouter } from "next/router";
-import { useEffect } from "react";
-import { useInView } from "react-intersection-observer";
 
-import { useNotes } from "@/hooks/notes/useNotes";
+import { useInfiniteNotes } from "@/hooks/notes/useInfiniteNotes";
 import Goal from "@/views/notes/goal/Goal";
 import NoteList from "@/views/notes/note-list/NoteList";
 
@@ -10,13 +8,7 @@ export default function NotesPage() {
   const {
     query: { goalId },
   } = useRouter();
-  const { ref: inViewRef, inView } = useInView();
-  const { data, fetchNextPage, hasNextPage } = useNotes(Number(goalId));
-
-  useEffect(() => {
-    // 스크롤 감지 블럭이 화면에 들어오고 다음페이지가 존재하는 경우에 데이터 더 가져오기
-    if (inView && hasNextPage) fetchNextPage();
-  }, [inView, hasNextPage]);
+  const { data, inViewRef } = useInfiniteNotes(Number(goalId));
 
   return (
     <div className="min-h-screen bg-slate-100">
